requirements.txt
tensorflow
rerun-sdk>=0.12.0a4
